+++ src/clj/rems/api/services/licenses.clj
@@ -11,20 +11,20 @@
   (:import [java.io FileInputStream ByteArrayOutputStream]))
 
 (defn create-license! [{:keys [licensetype organization localizations]} user-id]
-  (or (util/forbidden-organization-error organization)
-      (let [license (db/create-license! {:owneruserid user-id
-                                         :modifieruserid user-id
-                                         :organization (or organization "")
-                                         :type licensetype})
-            licid (:id license)]
-        (doseq [[langcode localization] localizations]
-          (db/create-license-localization! {:licid licid
-                                            :langcode (name langcode)
-                                            :title (:title localization)
-                                            :textcontent (:textcontent localization)
-                                            :attachmentId (:attachment-id localization)}))
-        {:success (not (nil? licid))
-         :id licid})))
+  (util/check-allowed-organization! organization)
+  (let [license (db/create-license! {:owneruserid user-id
+                                     :modifieruserid user-id
+                                     :organization (or organization "")
+                                     :type licensetype})
+        licid (:id license)]
+    (doseq [[langcode localization] localizations]
+      (db/create-license-localization! {:licid licid
+                                        :langcode (name langcode)
+                                        :title (:title localization)
+                                        :textcontent (:textcontent localization)
+                                        :attachmentId (:attachment-id localization)}))
+    {:success (not (nil? licid))
+     :id licid}))
 
 (defn create-license-attachment! [{:keys [tempfile filename content-type]} user-id]
   (attachments/check-attachment-content-type content-type)