+++ test/clj/rems/api/test_resources.clj
@@ -136,13 +136,18 @@
                 (is (= [licid-org1] (map :id (:licenses resource))))))))
 
         (testing "with incorrect organization"
-          (let [result (create-resource "organization-owner1" "organization2" licid-org2)]
-            (is (false? (:success result)))))
+          (let [response (-> (request :post "/api/resources/create")
+                             (authenticate api-key "organization-owner1")
+                             (json-body {:resid resid
+                                         :organization "organization2"
+                                         :licenses [licid-org2]})
+                             handler)]
+            (is (response-is-forbidden? response))
+            (is (= "no access to organization \"organization2\"" (read-body response)))))
 
-        (testing "with mismatched organizations"
+        (testing "with license from other organization"
           (let [result (create-resource "organization-owner1" "organization1" licid-org2)]
-            (is (false? (:success result)))
-            (is (= ["t.administration.errors/organization-mismatch"] (map :type (:errors result)))))))
+            (is (true? (:success result))))))
 
       (testing "create as owner"
         (let [result (create-resource "owner" "organization1" licid-org1)
@@ -168,8 +173,7 @@
               (is (true? (:success result))))))
         (testing "with mismatched organizations"
           (let [result (create-resource "owner" "organization1" licid-org1 licid-org2)]
-            (is (false? (:success result)))
-            (is (= ["t.administration.errors/organization-mismatch"] (map :type (:errors result))))))))))
+            (is (true? (:success result)))))))))
 
 (deftest resources-api-filtering-test
   (let [api-key "42"